+++ src/Pages/App.scss
@@ -1,0 +1,12 @@
+@import '../assets/styles/base/resets';
+@import '../assets/styles/base/variables';
+@import '../assets/styles/base/fonts';
+@import '../assets/styles/base/spacing';
+@import '../assets/styles/base/common';
+
+@import '../assets/styles/components/form';
+@import '../assets/styles/components/checkbox';
+@import '../assets/styles/components/textField';
+@import '../assets/styles/components/spinner';
+@import '../assets/styles/components/switch';
+@import '../assets/styles/components/tooltip';
