+++ src/Pages/Onboarding/steps/backup.scss
@@ -1,4 +1,4 @@
-@import '../../../base/variables';
+@import '../../../assets/styles/base/variables';
 
 .step-block.backup {
   width: 40%;
