--- src/assets/styles/App.scss
@@ -1,12 +1,0 @@
-@import 'base/resets';
-@import 'base/variables';
-@import 'base/fonts';
-@import 'base/spacing';
-@import 'base/common';
-
-@import 'components/form';
-@import 'components/checkbox';
-@import 'components/textField';
-@import 'components/spinner';
-@import 'components/switch';
-@import 'components/tooltip';
