--- src/assets/styles/pages/sessionsList.scss
@@ -1,19 +1,0 @@
-.table-container {
-  height: 60%;
-}
-
-.sessions {
-  display: flex;
-  height: 100%;
-
-  &--content {
-    width: 70%;
-    padding: 0 25px 100px 0;
-  }
-
-  //side block
-  &--side {
-    width: 30%;
-    background: #FFFFFF;
-  }
-}
